BAW_back/requirements.txt
Django==5.2.9
djangorestframework==3.16.1
django-cors-headers==4.9.0
djangorestframework-simplejwt==5.3.2
django-filter==24.1
pillow==10.2.0
python-decouple==3.8
